:truck: Refactor #21: 스페이스 화면 - 라벨 탭 관련 presentation 계층 폴더 이동
